backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.16
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Authentication & Security
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
passlib[bcrypt]>=1.7.4

# Pydantic & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# QR Code Generation
qrcode[pil]>=7.4.2
pillow>=10.2.0

# HTTP Client (for OAuth)
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
